refactor: move the files in to the layout folder
